Consolidate pthread functions into a single file
