--- angr/procedures/posix/pthread_mutex_lock.py
@@ -1,9 +1,0 @@
-import angr
-
-######################################
-# Doing nothing
-######################################
-
-class pthread_mutex_lock(angr.SimProcedure):
-    def run(self):
-        _ = self.arg(0)
