--- angr/procedures/posix/pthread_mutex_unlock.py
@@ -1,9 +1,0 @@
-import angr
-
-######################################
-# Doing nothing
-######################################
-
-class pthread_mutex_unlock(angr.SimProcedure):
-    def run(self):
-        _ = self.arg(0)
